Zero initialize SPI structs and add override_freq_hz (#20)
ESP-IDF 5.5.0 introduced a new field `override_freq_hz` to `spi_transaction_t`. Since the struct is not zero initialized and the new field is not set, it contained undefined data, which caused the SPI clock to peg at the APB clock frequency.

Explicitly zero initialize the SPI structs to initalize the new field and also future fields with zero. Also add a comment hinting at the new `override_freq_hz`.